- move buttons from third-party-plugin into core plugin under registry
